requirements.txt
amqp==5.3.1
asgiref==3.11.0
beautifulsoup4==4.14.3
billiard==4.2.4
cachetools==6.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
greenlet==3.3.0
httplib2==0.31.0
idna==3.11
kombu==5.6.2
lxml==6.0.2
numpy==2.4.0
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
phonenumbers==9.0.21
playwright==1.57.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pyee==13.0.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
RapidFuzz==3.14.3
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sentry-sdk==2.48.0
six==1.17.0
soupsieve==2.8.1
sqlparse==0.5.5
structlog==25.5.0
tenacity==9.1.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
gunicorn==24.1.1
whitenoise==6.11.0